refactor(components): [time-picker] constants (#8170)
* refactor(components): [time-picker] constants

* Relocate the constants and utils.

* chore: remove expression `instanceof`

* chore: refine code
