+++ packages/components/time-picker/index.ts
@@ -5,8 +5,8 @@ import TimePickPanel from './src/time-picker-com/panel-time-pick.vue'
 import type { App } from 'vue'
 import type { SFCWithInstall } from '@element-plus/utils'
 
-export * from './src/common/date-utils'
-export * from './src/common/constant'
+export * from './src/utils'
+export * from './src/constants'
 export * from './src/common/props'
 
 const _TimePicker = TimePicker as SFCWithInstall<typeof TimePicker>
